Add functions for DOM Manipulation
Added `buttonFunction` that connects monitors a `click` event on the button and triggers `showMessage()`. 

`showMessage()` takes a function as a parameter and writes the output to a DOM element with ID `quote`. 

Known issues:
`showMessage()` creates returns a `TypeError` when trying to access properties of `randomMessage()`. 
The remainder of the function works.